% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{signed_area}
\alias{signed_area}
\alias{unsigned_area}
\alias{signed_area_geodesic}
\alias{unsigned_area_geodesic}
\title{Calculate the area of a polygon}
\usage{
signed_area(x)

unsigned_area(x)

signed_area_geodesic(x)

unsigned_area_geodesic(x)
}
\arguments{
\item{x}{an object of class \code{rsgeo}}
}
\description{
Functions to calculate different types of area for polygons.
}
