% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_rsgeo}
\alias{as_rsgeo}
\title{Convert to an \code{rsgeo} vector}
\usage{
as_rsgeo(x)
}
\arguments{
\item{x}{a geometry vector}
}
\value{
an object of class \code{rsgeo}
}
\description{
Given an vector of geometries, cast it as an \code{rsgeo} class object.
}
\examples{
\dontshow{if (rlang::is_installed("sf")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- sf::st_sfc(sf::st_point(c(0,0)))
as_rsgeo(x)
\dontshow{\}) # examplesIf}
}
