% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bearing}
\alias{bearing}
\alias{bearings}
\title{Calculate Bearing}
\usage{
bearing(x, y)

bearings(x, y)
}
\arguments{
\item{x}{an object of class \code{point}}

\item{y}{for \code{bearing()} an object of class \code{point}. For \code{bearings()} an object of class \code{rs_POINT}}
}
\value{
A vector of doubles of the calculated bearing for between x and y
}
\description{
Calculate Bearing
}
