% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bounding_boxes}
\alias{bounding_boxes}
\alias{bounding_rect}
\alias{minimum_rotated_rect}
\alias{convex_hull}
\alias{concave_hull}
\alias{extreme_coords}
\alias{bounding_box}
\title{Compute Geometric Boundaries}
\usage{
bounding_boxes(x)

bounding_rect(x)

minimum_rotated_rect(x)

convex_hull(x)

concave_hull(x, concavity)

extreme_coords(x)

bounding_box(x)
}
\arguments{
\item{x}{a rust geometry either a scalar or a vector for functions ending in \code{s}. See "Details" for more.}
}
\description{
Compute Geometric Boundaries
}
\details{
\itemize{
\item \code{bounding_box()} returns a named vector of xmin, ymin, xmax, and ymax
\item \code{bounding_boxes()} returns a list of bounding
\item \code{bounding_rectangle()} returns a polygon of the bounding rectangle
\item \code{convex_hull()} returns a polygon of the convex hull for each geometry
\item \code{concave_hull()} returns a polygon of the specified concavity for each geometry
\item \code{extreme_coords()} returns the extreme coordinates of a geometry as a named vector of xmin, ymin, xmax, and ymax where each element is a \code{Point} geometry of the extreme value
\item \code{minimum_rotated_rect()} returns the minimum rotated rectangle covering a geometry
Note that if you want a convex or concave hull over an entire vector of geometries
you must first union or combine them.
}
}
