% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{centroids}
\alias{centroids}
\title{Extract Centroids}
\usage{
centroids(x)
}
\arguments{
\item{x}{an object of class \code{rsgeo}}
}
\value{
an object of class \code{rs_POINT}
}
\description{
Given a vector of geometries, extract their centroids.
}
\examples{
lns <- geom_linestring(1:100, runif(100, -10, 10), rep.int(1:5, 20))
centroids(lns)
}
