% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R, R/extendr-wrappers.R
\name{combine_geoms}
\alias{combine_geoms}
\alias{union_geoms}
\title{Combine geometries}
\usage{
combine_geoms(x)

union_geoms(x)
}
\arguments{
\item{x}{a vector of geometries}
}
\description{
Combine geometries
}
