% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{euclidean_distance}
\alias{euclidean_distance}
\alias{euclidean_distance_pairwise}
\alias{euclidean_distance_matrix}
\alias{haversine_distance}
\alias{haversine_distances}
\alias{haversine_distance_matrix}
\alias{geodesic_distance}
\alias{geodesic_distances}
\alias{geodesic_distance_matrix}
\alias{vincenty_distance}
\alias{vincenty_distances}
\alias{vincenty_distance_matrix}
\title{Distance calculations}
\usage{
euclidean_distance(x, y)

euclidean_distance_pairwise(x, y)

euclidean_distance_matrix(x, y)

haversine_distance(x, y)

haversine_distances(x, y)

haversine_distance_matrix(x, y)

geodesic_distance(x, y)

geodesic_distances(x, y)

geodesic_distance_matrix(x, y)

vincenty_distance(x, y)

vincenty_distances(x, y)

vincenty_distance_matrix(x, y)
}
\arguments{
\item{x}{a single \code{point} or list of points \code{rs_POINT}}

\item{y}{a single \code{point} or list of points \code{rs_POINT}}
}
\description{
Distance calculations
}
