% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{explode_lines}
\alias{explode_lines}
\title{Explode Lines}
\usage{
explode_lines(x)
}
\arguments{
\item{x}{an object of class \code{rs_LINESTRING} or \code{rs_MULTILINESTRING}}
}
\value{
an object of class \code{rs_LINESTRING}
}
\description{
Given a LineString or MultiLineString, expand the geometry into each and
every component Line.
}
\details{
A \code{LineString} is composed of one or more \code{Line}s. A Line is a connected
by a start and end coordinate only.
}
\examples{
x <- geom_linestring(1:10, 10:1)
length(x)
explode_lines(x)
}
