% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_geoms}
\alias{flatten_geoms}
\title{Flatten a list of rsgeo vectors}
\usage{
flatten_geoms(x)
}
\arguments{
\item{x}{list object where each element is an object of class \code{rsgeo}}
}
\value{
Returns an object of class \code{rsgeo}
}
\description{
Flatten a list of rsgeo vectors
}
\examples{
pnts <- replicate(
  10,
  geom_point(runif(1), runif(1)),
  simplify = FALSE
)

flatten_geoms(pnts)

}
