% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{frechet_distance}
\alias{frechet_distance}
\title{Calculate Frechet Distance}
\usage{
frechet_distance(x, y)
}
\arguments{
\item{x}{an object of class \code{rs_LINESTRING}}

\item{y}{an object of class \code{rs_LINESTRING}}
}
\value{
A numeric vector
}
\description{
Given two LineStrings compare thier similarity
by calculating the Fréchet distance.
}
\examples{
x <- geom_linestring(1:10, runif(10, -1, 1))
y <- geom_linestring(1:10, runif(10, -3, 3))
frechet_distance(x, y)
}
