% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{geom_points_matrix}
\alias{geom_points_matrix}
\title{Create a list of points
Given a matrix of x, y coordinates, create a list of points}
\usage{
geom_points_matrix(x)
}
\description{
Create a list of points
Given a matrix of x, y coordinates, create a list of points
}
