% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{geom_point}
\alias{geom_point}
\alias{geom_points}
\alias{geom_points_xy}
\alias{geom_points_matrix}
\alias{geom_multipoint}
\alias{geom_multipoints}
\alias{geom_polygon}
\alias{geom_polygons}
\alias{geom_multipolygon}
\alias{geom_multipolygons}
\alias{geom_linestring}
\alias{geom_linestrings}
\alias{geom_multilinestring}
\alias{geom_multilinestrings}
\title{Create geometry}
\usage{
geom_point(x, y)

geom_points(x)

geom_points_xy(x, y)

geom_points_matrix(x)

geom_multipoint(x)

geom_multipoints(x)

geom_polygon(x)

geom_polygons(x)

geom_multipolygon(x)

geom_multipolygons(x)

geom_linestring(x)

geom_linestrings(x)

geom_multilinestring(x)

geom_multilinestrings(x)
}
\description{
Create geometry

Create a list of points
Given a matrix of x, y coordinates, create a list of points
}
