% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{plot.rsgeo}
\alias{plot.rsgeo}
\title{Plot Geometries}
\usage{
\method{plot}{rsgeo}(x, ...)
}
\arguments{
\item{x}{an object of class \code{rsgeo}}

\item{...}{arguments passed to \code{wk::wk_plot()}}
}
\description{
Plot Geometries
}
\details{
Plotting geometries utilizes \code{wk::wk_plot()}. The rust geometries are
handled by first converting to an \code{sfc} object in the \code{wk::wk_handle()}
method thus requiring both packages for plotting.
}
