% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{contains}
\alias{contains}
\alias{contains_sparse}
\alias{contains_pairwise}
\alias{intersects_sparse}
\alias{intersects}
\alias{intersects_pairwise}
\alias{within}
\alias{within_sparse}
\alias{within_pairwise}
\title{Spatial Predicates}
\usage{
contains(x, y)

contains_sparse(x, y)

contains_pairwise(x, y)

intersects_sparse(x, y)

intersects(x, y)

intersects_pairwise(x, y)

within(x, y)

within_sparse(x, y)

within_pairwise(x, y)
}
\description{
Spatial Predicates
}
