% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.point}
\alias{print.point}
\alias{print.multipoint}
\alias{print.polygon}
\alias{print.multipolygon}
\alias{print.linestring}
\alias{print.multilinestring}
\title{Printing and formatting}
\usage{
\method{print}{point}(x, width = NULL)

\method{print}{multipoint}(x, width = NULL)

\method{print}{polygon}(x, width = NULL)

\method{print}{multipolygon}(x, width = NULL)

\method{print}{linestring}(x, width = NULL)

\method{print}{multilinestring}(x, width = NULL)
}
\description{
Printing and formatting
}
