% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{read_geojson}
\alias{read_geojson}
\title{Read GeoJSON file}
\usage{
read_geojson(filepath)
}
\arguments{
\item{filepath}{the file path to be read}
}
\description{
Read a GeoJson file using the \code{geojson} crate. This function will crash on
geojson files ~ 50mb +. Working on it though....
}
\details{
To improve reading speed set \code{--max-ppsize=500000} when creating an R session.
}
