% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_points}
\alias{rs_points}
\title{Create a list of points}
\usage{
rs_points(x)
}
\description{
Given a matrix of x, y coordinates, create a list of points
uses the \code{mat_to_rs()} helper to convert an RMatrix to ndarray::Array2
}
