% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{rsgeo-package}
\alias{rsgeo}
\alias{rsgeo-package}
\title{rsgeo: An Interface to Rust's 'geo' Library}
\description{
An R interface to the GeoRust crates 'geo' and 'geo-types' providing access to geometry primitives and algorithms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JosiahParry/rsgeo}
  \item \url{https://josiahparry.r-universe.dev/rsgeo}
  \item \url{https://rsgeo.josiahparry.com/}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

}
