% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rstar_rtree}
\alias{rstar_rtree}
\title{Create an rstar RTree from a vector of geometry}
\usage{
rstar_rtree(x)
}
\arguments{
\item{x}{a vector of rust geometry}
}
\description{
Create an rstar RTree from a vector of geometry
}
