% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{simplify_geom}
\alias{simplify_geom}
\alias{simplify_geoms}
\alias{simplify_vw_geom}
\alias{simplify_vw_geoms}
\title{Simplfiy Geometries}
\usage{
simplify_geom(x, epsilon)

simplify_geoms(x, epsilon)

simplify_vw_geom(x, epsilon)

simplify_vw_geoms(x, epsilon)
}
\description{
Simplfiy Geometries
}
