% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{intersects_sparse}
\alias{intersects_sparse}
\alias{intersects_pairwise}
\alias{contains_sparse}
\alias{contains_pairwise}
\alias{within_sparse}
\alias{within_pairwise}
\title{Binary Predicates}
\usage{
intersects_sparse(x, y)

intersects_pairwise(x, y)

contains_sparse(x, y)

contains_pairwise(x, y)

within_sparse(x, y)

within_pairwise(x, y)
}
\arguments{
\item{x}{an object of class \code{rsgeo}}

\item{y}{an object of class \code{rsgeo}}
}
\description{
Functions to ascertain the binary relationship between
two geometry vectors. Binary predicates are provided both pairwise
as a sparse matrix.
}
