% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wkb_from_geom}
\alias{wkb_from_geom}
\alias{wkb_from_geoms}
\alias{wkb_to_geom}
\alias{wkb_to_geoms}
\title{Convert to and from WKB}
\usage{
wkb_from_geom(x)

wkb_from_geoms(x)

wkb_to_geom(x)

wkb_to_geoms(x)
}
\description{
Convert to and from WKB
}
