% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{wkt_to_geom}
\alias{wkt_to_geom}
\alias{wkt_from_geom}
\alias{wkt_to_geoms}
\alias{wkt_from_geoms}
\title{WKT translation}
\usage{
wkt_to_geom(x)

wkt_from_geom(x)

wkt_to_geoms(x)

wkt_from_geoms(x)
}
\description{
WKT translation
}
