% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bounding_box}
\alias{bounding_box}
\alias{bounding_rectangle}
\alias{bounding_rectangles}
\alias{concave_hull}
\alias{concave_hulls}
\alias{convex_hull}
\alias{convex_hulls}
\title{Compute Geometric Boundaries}
\usage{
bounding_box(x)

bounding_rectangle(x)

bounding_rectangles(x)

concave_hull(x, concavity)

concave_hulls(x, concavity)

convex_hull(x)

convex_hulls(x)
}
\arguments{
\item{x}{a rust geometry either a scalar or a vector for functions ending in \code{s}. See "Details" for more.}
}
\description{
Compute Geometric Boundaries
}
\details{
\itemize{
\item \code{bounding_box()} returns a named list of x and y maximums and minimums
\item \code{bounding_rectangle()} returns a polygon of the bounding rectangle
\item \code{convex_hull()} returns a polygon of the convex hull
\item \code{concave_hull()} returns a polygon of the specified concavity
}

Each function, with the exception of \code{bounding_box()} has a plural version ending
with an \code{s} which is vectorized over \code{x}.
}
