% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{centroids}
\alias{centroids}
\title{Extract Centroids}
\usage{
centroids(x)
}
\arguments{
\item{x}{an object of class \code{rsgeo}}
}
\description{
Given a vector of geometries, extract their centroids.
}
