% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{closest_point}
\alias{closest_point}
\title{Find the closest point}
\usage{
closest_point(x, y)
}
\arguments{
\item{x}{a single geometry}

\item{y}{a \code{point}}
}
\description{
Find the closest point
}
