% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{line_interpolate_point}
\alias{line_interpolate_point}
\title{Interpolate a Point on a LineString}
\usage{
line_interpolate_point(x, fraction)
}
\arguments{
\item{x}{an object of class \code{rs_LINESTRING}}

\item{y}{a numeric vector of the same length as \code{x} or length 1}
}
\value{
An object of class \code{rs_POINT}
}
\description{
Finds the point that lies a given fraction along a line.
}
\examples{
x <- geom_linestring(c(-1, 0, 0), c(0, 0, 1))
line_interpolate_point(x, 0.5)
}
